backend/requirements.txt
attrs==23.1.0
click==8.1.3
flasgger==0.9.5
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mistune==2.0.5
psycopg2-binary==2.9.6
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==0.21.1
PyYAML==6.0
six==1.16.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
